requirements.txt
fastapi
uvicorn
langchain
unstructured
chromadb
tesseract
poppler-utils
unstructured[all-docs]
pillow
lxml
chromadb
tiktoken
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
fastapi
python-multipart
uvicorn
unstructured[all-docs]
pillow
lxml
chromadb
tiktoken
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv